Initial draft amendment of repository for use in workshop format
