+++ .github/workflows/deploy.yml
@@ -34,11 +34,11 @@ jobs:
     # Build the book
     - name: Build the book
       run: |
-        jupyter-book build data-science-for-esm
+        jupyter-book build open-tyndp-workshops
 
     # Deploy the book's HTML to gh-pages branch
     - name: GitHub Pages action
       uses: peaceiris/actions-gh-pages@v3.6.1
       with:
         github_token: ${{ secrets.GITHUB_TOKEN }}
-        publish_dir: data-science-for-esm/_build/html
+        publish_dir: open-tyndp-workshops/_build/html
