+++ .gitignore
@@ -1,12 +1,12 @@
 .ipynb-checkpoints
-data-science-for-esm/data
+open-tyndp-workshops/data
 *.geojson
 *.nc
 *.tiff
 *.tif
 *.html
 *.png
-data-science-for-esm/tmp/
-data-science-for-esm/_build/
-data-science-for-esm/tmp.csv
+open-tyndp-workshops/tmp/
+open-tyndp-workshops/_build/
+open-tyndp-workshops/tmp.csv
 *cleared.ipynb
