+++ README.md
@@ -1,20 +1,22 @@
-# Data Science for Energy System Modelling
+# Open-TYNDP Workshops
 
-Course at TU Berlin to learn energy system modelling with data.
+Series of hands-on workshops facilitated by Open Energy Transition (OET) to accompany the development of an open-source energy modeling tool for the TYNDP.
+With every workshop a new notebook will be added to the repository investigating different functionalities of PyPSA 
+and implementations for the Open-TYNDP workflow.
 
 ## Usage
 
 ### Building the book
 
-If you'd like to develop and/or build the Data Science for Energy System Modelling book, you should:
+If you'd like to develop and/or build the Open-TYNDP Workshops book, you should:
 
 1. Clone this repository
 2. Run `pip install -r requirements.txt` (it is recommended you do this within a virtual environment)
-3. (Optional) Edit the books source files located in the `data-science-for-esm/` directory
-4. Run `jupyter-book clean data-science-for-esm/` to remove any existing builds
-5. Run `jupyter-book build data-science-for-esm/`
+3. (Optional) Edit the books source files located in the `open-tyndp-workshops/` directory
+4. Run `jupyter-book clean open-tyndp-workshops/` to remove any existing builds
+5. Run `jupyter-book build open-tyndp-workshops/`
 
-A fully-rendered HTML version of the book will be built in `data-science-for-esm/_build/html/`.
+A fully-rendered HTML version of the book will be built in `open-tyndp-workshops/_build/html/`.
 
 ### Hosting the book
 
@@ -28,4 +30,4 @@ We welcome and recognize all contributions. You can see a list of current contri
 
 ## Credits
 
-This project is created using the excellent open source [Jupyter Book project](https://jupyterbook.org/) and the [executablebooks/cookiecutter-jupyter-book template](https://github.com/executablebooks/cookiecutter-jupyter-book).
+This project is created by forking of [Fabian Neumann](https://github.com/fneum)'s excellent open-source course [Data Science for Energy System Modelling](https://github.com/fneum/data-science-for-esm) which uses the open source [Jupyter Book project](https://jupyterbook.org/) and the [executablebooks/cookiecutter-jupyter-book template](https://github.com/executablebooks/cookiecutter-jupyter-book).
