--- data-science-for-esm/_toc.yml
@@ -1,23 +1,0 @@
-# Table of contents
-# Learn more at https://jupyterbook.org/customize/toc.html
-
-format: jb-book
-root: intro
-chapters:
-- file: 01-workshop-python
-- file: 02-workshop-numpy
-- file: 03-workshop-pandas
-- file: 04-workshop-geopandas
-- file: 06-workshop-atlite
-- file: 05-workshop-pysheds
-- file: 07-workshop-networkx
-- file: 14-workshop-linopy
-- file: 09-workshop-pypsa
-- file: 10-workshop-pypsa-cem
-- file: 11-workshop-groupwork
-- file: 12-workshop-pypsa-sector-coupling
-- file: 13-workshop-interactive-visualisation
-- file: 08-workshop-pyomo
-
-- file: oet-extension/00-build_your_own_content.md
-- file: oet-extension/00-reference.md
