--- data-science-for-esm/oet-extension/00-reference.md
@@ -1,79 +1,0 @@
-<!-- Relevant References for further reading -->
-References
-=============================
-
-::::{tab-set}
-
-:::{tab-item} Jupyter-book
-1. https://jupyterbook.org/en/stable/start/create.html
-1. https://jupyterbook.org/en/stable/structure/toc.html
-1. https://jupyterbook.org/en/stable/structure/configure.html
-1. https://jupyterbook.org/en/stable/customize/config.html
-:::
-
-:::{tab-item} Markdown
-1. https://jupyterbook.org/en/stable/reference/cheatsheet.html
-1. https://myst-parser.readthedocs.io/en/latest/syntax/admonitions.html
-1. https://www.markdownguide.org/extended-syntax/
-:::
-
-:::{tab-item} PyPSA-Earth
-* https://pypsa-eur.readthedocs.io/en/latest/introduction.html
-:::
-
-:::{tab-item} Git
-1. https://www.atlassian.com/git
-1. [https://videotutorials.notion.site/Introduction-to-Git](https://videotutorials.notion.site/Introduction-to-Git-ac396a0697704709a12b6a0e545db049)
-1. https://git-scm.com/
-::::
-
-::::{tab-set}
-
-:::{tab-item} OET
-1. https://openenergytransition.org/about-us.html#theory-of-change
-1. https://open-energy-transition.github.io/handbook/
-:::
-
-:::{tab-item} Data Science for ESM
-* https://www.tu.berlin/en/ensys
-<!-- 1. https://moseskonto.tu-berlin.de/moses/modultransfersystem/bolognamodule/beschreibung/anzeigen.html;jsessionid=DQfixqzzpn1XIg5N1GG7S9um4EDykZn99AHmH6Fj.moseskonto?number=31027&version=1&sprache=2 -->
-<!-- 1. https://fneum.github.io/data-science-for-esm/09-workshop-pypsa.html -->
-:::
-
-:::{tab-item} Dr. Fabian Neumann
-* https://www.neumann.fyi/
-:::
-
-::::
-
-<!-- 1. Jupyter-book
-    * https://jupyterbook.org/en/stable/start/build.html
-    * https://jupyterbook.org/en/stable/structure/toc.html
-1. MyST
-    * https://jupyterbook.org/en/stable/reference/cheatsheet.html
-    * https://myst-parser.readthedocs.io/en/latest/syntax/admonitions.html
-    * https://executablebooks.github.io/cookiecutter-jupyter-book/notebooks.html
-1. PyPSA-Earth:
-    * https://pypsa-eur.readthedocs.io/en/latest/introduction.html
-1. Open Energy Transition
-    * https://openenergytransition.org/about-us.html#theory-of-change
-    * https://open-energy-transition.github.io/handbook/
-1. Data Science for Energy System Modelling
-    * https://www.tu.berlin/ensys
-    * https://fneum.github.io/data-science-for-esm/09-workshop-pypsa.html
-    * https://moseskonto.tu-berlin.de/moses/modultransfersystem/bolognamodule/beschreibung/anzeigen.html;jsessionid=DQfixqzzpn1XIg5N1GG7S9um4EDykZn99AHmH6Fj.moseskonto?number=31027&version=1&sprache=2
-1. Dr. Fabian Neumann
-    * https://www.neumann.fyi/ -->
-
-
-<!-- Citation
---------- -->
-<!-- <mark style="background-color: peachpuff; color: black"> -->
-<!-- *Kindly [reference][OET] us, in case using our materials, e.g. the workflow presented here.
-This way, your success is our success also.* -->
-\
-\
-_Best,
-<br />The [OET Team][OET]_
-
-[OET]:  https://openenergytransition.org/about-us.html#team
