+++ open-tyndp-workshops/_config.yml
@@ -1,0 +1,53 @@
+#######################################################################################
+# A default configuration that will be loaded for all jupyter books
+# See the documentation for help and more options:
+# https://jupyterbook.org/customize/config.html
+
+#######################################################################################
+# Book settings
+title: Open-TYNDP Workshops # The title of the book. Will be placed in the left navbar.
+author: Open Energy Transition # The author of the book
+copyright: 2025                                     # Copyright year to be placed in the footer
+logo: logo_merged.png                           # A path to the book logo
+
+# Force re-execution of notebooks on each build.
+# See https://jupyterbook.org/content/execute.html
+execute:
+  execute_notebooks: force
+
+# Define the name of the latex output file for PDF builds
+latex:
+  latex_documents:
+    targetname: book.tex
+
+# Add a bibtex file so that we can create citations
+bibtex_bibfiles:
+- references.bib
+
+# Information about where the book exists on the web
+repository:
+  url: https://github.com/open-energy-transition/open-tyndp-workshops  # Online location of your book
+  path_to_book: open-tyndp-workshops  # Optional path to your book, relative to the repository root
+  branch: main  # Which branch of the repository should be used when creating links (optional)
+
+# Add GitHub buttons to your book
+# See https://jupyterbook.org/customize/config.html#add-a-link-to-your-repository
+html:
+  use_issues_button: true
+  use_repository_button: true
+
+launch_buttons:
+  colab_url: "https://colab.research.google.com"
+
+parse:
+  myst_enable_extensions:
+  - amsmath
+  - dollarmath
+  - linkify
+  - colon_fence
+  - substitution
+
+sphinx:
+  config:
+    html_js_files:
+    - https://cdnjs.cloudflare.com/ajax/libs/require.js/2.3.4/require.min.js
