+++ open-tyndp-workshops/references.bib
@@ -1,0 +1,4 @@
+---
+---
+
+@article{Hofmann2021, doi = {10.21105/joss.03294}, url = {https://doi.org/10.21105/joss.03294}, year = {2021}, publisher = {The Open Journal}, volume = {6}, number = {62}, pages = {3294}, author = {Fabian Hofmann and Johannes Hampp and Fabian Neumann and Tom Brown and Jonas Hörsch}, title = {atlite: A Lightweight Python Package for Calculating Renewable Power Potentials and Time Series}, journal = {Journal of Open Source Software} }
